requirements.txt
apache-airflow==2.0.2
apache-airflow-providers-slack==3.0.0
freezegun==1.1.0 
requests==2.22.0
